Updates code so that we display new users in a different way
The line annotations looked a bit naff, and it didn't seem worth the
effort in trying to make that kind of annotation look nicer - it felt like
it would be trying to force D3 to do something it's not designed for.
Instead, I now just fade in a circle below the chart to indicate a new
user event.