carousel caption is now clickable.
The HTML included the carousel caption div within the anchor, but
the div wasn't included within the anchor tag. Removing the
include of social.html seemed to fix this.

I'm not 100% sure why, but I suspect that Jekyll may be strictly
enforcing the HTML5 spec which states that links can wrap other
elements "as there is no interactive content within (e.g.
buttons or other links)". Since social.html includes links, I suspect
that Jekyll was treating the anchor tag as unclosed and inserting
a close anchor tag directly after the image - causing the caption
to be excluded from the anchor and thus end up non-clickable.