Cite is now visible below carousel item.
The cite was positioned absolutely with a negative offset, meaning
it overflowed out of it's containing div. Since the carousel
is set to hide it's overflow, the citation was being clipped.

To fix this, I added a 'carousel-caption-container' class with
a relative position to ensure that the absolute 'carousel-caption'
positions relative to it.

The cite no longer needs to be positioned relatively, it can simply
float right with the equivalent margins.