Small screen carousel displays to the edge of screen.
The carousel now displays to the edge of the screen with a margin
of 20px at the edges of it's content. This means the carousel can
expand all the way to the edge of the screen but still retain it's
original white-space margin whilst it is at the end of it's content.

To do this, I've had to have the small carousel overflow the
pages parent container which was adding a 20px gutter to the edges
of the page.