Refactors duplicate carousel code.
The code for creating a column on the carousel is now in its
own include. Since they do vary, there are various parameters that
are passed to modify the content.

I think this is a good balance as the parameters are pretty self
explanatory, and are things that you could easily imagine changing
on the columns.