Fixes issue whereby lazily loaded posts have incorrect padding.
Paginated posts were inset an extra 20px than the original page
of posts. This is because jScroll duplicates the markup of a given
element, and it was duplicating the markup of our 'content' div
which holds the blog content. This 'content' div has 20px padding,
so the duplicated content thus had 40px padding.

To fix this I wrapped added a wrapper div around 'content' and
added the padding to that instead.