Social media icons on post summaries are now grey.
Explicitly specify that post summaries should NOT display their
social icons as black.

Since we weren't specifying the value of 'black', it was taking
on the value of the last featured blog post. With the addition
of the react blog post the last featured block post was set
black to true. Thus the social media icons were rendered black.