Re-adds search box functionality.
This was removed when the ScottLogic app.js script was removed.
We're now only adding back the bits that are relevant to the
ScottLogic blog.

We've left out:
- A scroll to top on page load, which was the motivation for the
change.
- the mobile search button functionality, there's no mobile
search button so thus no functional change here.
- The font sizing / re-sizing code, this was specific to elements
on the ScottLogic website, these elements don't exist on the blog
so thus no functional change here.