Colin's blog page now uses the featured_carousel.
His blog page simply passes the carousel a list of posts. In this
case, a list of posts with the tag 'featuredcolin'.

I've also removed a tag from a blog post. This post wasn't
being displayed on the original carousel as it was limited
to one page. It causes odd behaviour so I've removed this
post.