Updates blog index page featured articles
This area reacts responsively for four sizes:
~1200 ~ 1000 ~ 800 ~ 480

The design looks a little wrong at <480 and articles
should probably stack.

This is a rather large commit as it involves updating
from bootstrap 2 -> 3. I didn't think it was worth updating
the site and bringing it back into its original state before
implementing the design - it would be too much work.

Many of the other pages are broken at this point.