Adds responsive categories menu
This menu is shown either at a menu to the left side for md/lg
devices, or as a drop down menu for xs/sm devices.

There was a bit of rejigging to make the CSS currently used
for the main nav dropdown reusable in the categories menu
as before it was a little coupled to its HTML structure.