Removed 'shared assets concept' and instructions relating to git submodules
We were going to create a set of shared assets that would be used by the blogs, unleashed, events, etc ... however, the long-term plan for the non-blog sites is to move them to WordPress (the SL main website is being migrated to this platform). This makes the shared assets concept redundant.

Also removed the reference to git submodules, we're using npm instead.