Update markdown-spellcheck dependency (#705)
The markdown-spellcheck v1.3.1 has been released and includes the
required changes for relative spellcheck. The package.json dependency
that was previously linked to my fork tag has now been updated with the
official repo. The changes that are released in v.1.3.1 mean that more
of the blog posts content will be spellchecked that havn't been before
and require some updates. All updates are within this commit.