New post - bridging UI libraries (#1104)
I'd appreciate any feedback and sanity checking, as this was a bit of an exploration and experimentation -  rather than based on a solid proven thing done on a real project.

I've post-dated this as Wednesday to leave a bit of time after #1103. I'll swap the dates around if this happens to be approved before that one.

You'll see that the source branch is `gh-pages3`; this is because I had two posts on the go and both are on my main `gh-pages` branch.

Please add a direct link to your post here:

https://robatwilliams.github.io/blog/2018/12/12/bridging-ui-libraries-for-decoupled-long-lived-widely-reusable-components.html

Have you (please tick each box to show completion):

 - [x] Added your blog post to a single category?
 - [x] Added a brief summary for your post? Summaries should be roughly two sentences in length and give potential readers a good idea of the contents of your post.
 - [x] Checked that the build passes and spelling errors have been corrected?
 - [x] Ensured that your author profile contains a profile image (gravatar), and a brief description of yourself? (make it more interesting than just your job title!)
 - n/a Optimised any images in your post? They should be less than 100KBytes as a general guide.

Posts are approved based on their category. replace `(at)` with `@` to notify them!

 - Tech - Colin Eberhardt - @ColinEberhardt
 - Delivery - Matt Phillips - (at)scottlogicmphillips 
 - UX - Graham Odds - (at)godds
 - Careers - Rob Newsome - (at)rob-newsome
 - News - Patrick Bishop - (at)???
 - Testing - Laurence Pisani - (at)lpisani-SL